+++ src/RoutingMiddleware.php
@@ -1,0 +1,33 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Idiosyncratic\Http\Router;
+
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\MiddlewareInterface;
+use Psr\Http\Server\RequestHandlerInterface;
+use Throwable;
+
+class RoutingMiddleware implements MiddlewareInterface
+{
+    /** @var Router */
+    private $router;
+
+    public function __construct(Router $router)
+    {
+        $this->router = $router;
+    }
+
+    public function process(
+        ServerRequestInterface $request,
+        RequestHandlerInterface $handler
+    ) : ResponseInterface {
+        try {
+            return $this->router->dispatch($request);
+        } catch (Throwable $e) {
+            return $handler->handle($request);
+        }
+    }
+}